fix(antigravity): harden store discovery and schema probing
Probe steps.metadata as a column, not just the table, and re-probe while a
database lacks it so a store that gains step timestamps is re-read without a
restart. De-duplicate symlinked or repeated stores so conversations are not
counted twice. Surface a failed ~/.gemini listing through the read-failure
reporter instead of reading it as no usage. Add a real WAL-mode fixture test
and split the Antigravity test file by concern.
