+++ Tests/OpenUsageTests/AntigravityDbUsageScannerTests.swift
@@ -1,233 +1,6 @@
 import XCTest
 @testable import OpenUsage
 
-/// Synthetic protobuf fixtures follow FelixIsaac's original regression coverage in openusage#1058.
-private func antigravityVarint(_ value: UInt64) -> [UInt8] {
-    var remaining = value
-    var encoded: [UInt8] = []
-    repeat {
-        var byte = UInt8(remaining & 0x7f)
-        remaining >>= 7
-        if remaining != 0 { byte |= 0x80 }
-        encoded.append(byte)
-    } while remaining != 0
-    return encoded
-}
-
-private func antigravityVarintField(_ number: UInt32, _ value: UInt64) -> [UInt8] {
-    antigravityVarint(UInt64(number) << 3) + antigravityVarint(value)
-}
-
-private func antigravityBytesField(_ number: UInt32, _ bytes: [UInt8]) -> [UInt8] {
-    antigravityVarint(UInt64(number) << 3 | 2) + antigravityVarint(UInt64(bytes.count)) + bytes
-}
-
-private func antigravityHex(_ bytes: [UInt8]) -> String {
-    bytes.map { String(format: "%02x", $0) }.joined()
-}
-
-/// `steps.metadata`: field 1 is a Timestamp message whose field 1 is seconds.
-private func antigravityStepMetadata(timestamp: UInt64) -> [UInt8] {
-    antigravityBytesField(1, antigravityVarintField(1, timestamp))
-}
-
-private func antigravityGenerationBlob(
-    model: String?,
-    input: UInt64,
-    output: UInt64,
-    cacheRead: UInt64 = 0,
-    systemPrompt: UInt64 = 0,
-    label: String? = nil,
-    timestamp: UInt64?
-) -> [UInt8] {
-    let usage = antigravityVarintField(1, systemPrompt)
-        + antigravityVarintField(2, input)
-        + antigravityVarintField(3, output)
-        + antigravityVarintField(5, cacheRead)
-
-    var event: [UInt8] = []
-    if let model {
-        event += antigravityBytesField(19, Array(model.utf8))
-    }
-    event += antigravityBytesField(4, usage)
-    if let label {
-        event += antigravityBytesField(21, Array(label.utf8))
-    }
-
-    if let timestamp {
-        let wallClock = antigravityVarintField(1, timestamp)
-        event += antigravityBytesField(9, antigravityBytesField(4, wallClock))
-    }
-    return antigravityBytesField(1, event)
-}
-
-private struct AntigravityFixtureRow: Sendable {
-    var index: Int
-    var blob: [UInt8]?
-}
-
-private final class AntigravityFakeSQLite: SQLiteAccessing, @unchecked Sendable {
-    private let lock = NSLock()
-    private var rowsByPath: [String: [AntigravityFixtureRow]]
-    private var cursors: [Int] = []
-    let failingPaths: Set<String>
-    let cancellingPath: String?
-
-    init(rowsByPath: [String: [AntigravityFixtureRow]] = [:], failingPaths: Set<String> = [], cancellingPath: String? = nil) {
-        self.rowsByPath = rowsByPath
-        self.failingPaths = failingPaths
-        self.cancellingPath = cancellingPath
-    }
-
-    func queryValue(path: String, sql: String) throws -> String? {
-        lock.lock()
-        defer { lock.unlock() }
-
-        if failingPaths.contains(path) {
-            throw SQLiteError.queryFailed("database locked")
-        }
-        if path == cancellingPath { withUnsafeCurrentTask { $0?.cancel() } }
-        if sql == AntigravityDbUsageScanner.stepsTableProbeSQL { return nil }
-        let marker = "WHERE idx > "
-        guard let markerRange = sql.range(of: marker),
-              let cursor = Int(sql[markerRange.upperBound...].split(whereSeparator: \.isWhitespace).first ?? "")
-        else {
-            throw SQLiteError.queryFailed("missing batch cursor")
-        }
-        cursors.append(cursor)
-
-        let rows: [[String: Any]] = rowsByPath[path, default: []]
-            .filter { $0.index > cursor }
-            .sorted { $0.index < $1.index }
-            .prefix(AntigravityDbUsageScanner.batchSize)
-            .map { row in
-                let value: Any = row.blob.map(antigravityHex) ?? NSNull()
-                return ["index": row.index, "hex": value]
-            }
-
-        let json = try JSONSerialization.data(withJSONObject: rows)
-        return String(decoding: json, as: UTF8.self)
-    }
-
-    func append(_ row: AntigravityFixtureRow, to path: String) {
-        lock.withLock { rowsByPath[path, default: []].append(row) }
-    }
-
-    var queriedCursors: [Int] {
-        lock.withLock { cursors }
-    }
-
-    func execute(path: String, sql: String) throws {}
-}
-
-final class AntigravityProtoDecoderTests: XCTestCase {
-    func testDecodeVarintsAndRejectMalformedValues() {
-        XCTAssertEqual(AntigravityProtoDecoder.decodeVarint([0x05])?.value, 5)
-        XCTAssertEqual(AntigravityProtoDecoder.decodeVarint([0xac, 0x02])?.value, 300)
-        XCTAssertNil(AntigravityProtoDecoder.decodeVarint([0x80]))
-        XCTAssertNil(AntigravityProtoDecoder.decodeVarint(Array(repeating: 0xff, count: 11)))
-        XCTAssertNil(AntigravityProtoDecoder.decodeVarint(Array(repeating: 0xff, count: 9) + [0x02]))
-    }
-
-    func testMalformedLengthCannotOverflowOrTrap() {
-        let validField = antigravityVarintField(1, 42)
-        let oversizedLength = antigravityVarint(UInt64(2) << 3 | 2) + antigravityVarint(UInt64.max)
-        let fields = validField + oversizedLength
-
-        XCTAssertEqual(AntigravityProtoDecoder.varintField(1, in: fields), 42)
-        XCTAssertNil(AntigravityProtoDecoder.bytesField(2, in: fields))
-    }
-
-    func testExtractsCompleteGenerationRecord() throws {
-        let blob = antigravityGenerationBlob(
-            model: "gemini-3.1-pro-low",
-            input: 1_000,
-            output: 200,
-            cacheRead: 50,
-            systemPrompt: 1_132,
-            timestamp: 1_800_000_000
-        )
-        let event = try XCTUnwrap(AntigravityProtoDecoder.generationEvent(from: blob))
-
-        XCTAssertEqual(event.modelID, "gemini-3.1-pro-low")
-        XCTAssertNil(event.label)
-        XCTAssertEqual(event.inputTokens, 2_132)
-        XCTAssertEqual(event.outputTokens, 200)
-        XCTAssertEqual(event.cacheReadTokens, 50)
-        XCTAssertEqual(event.timestampSeconds, 1_800_000_000)
-    }
-
-    func testRejectsMissingTimestampZeroUsageAndUnrepresentableCounts() {
-        let missingTimestamp = antigravityGenerationBlob(model: "gemini-3.6-flash", input: 1, output: 0, timestamp: nil)
-        let zeroUsage = antigravityGenerationBlob(model: "gemini-3.6-flash", input: 0, output: 0, timestamp: 1_800_000_000)
-        let overflowingTokens = antigravityGenerationBlob(model: "gemini-3.6-flash", input: UInt64.max, output: 0, timestamp: 1_800_000_000)
-        let overflowingSystemPrompt = antigravityGenerationBlob(
-            model: "gemini-3.6-flash", input: UInt64(Int.max), output: 0,
-            systemPrompt: 1, timestamp: 1_800_000_000
-        )
-
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: missingTimestamp))
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: zeroUsage))
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: overflowingTokens))
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: overflowingSystemPrompt))
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: [0xff, 0xff, 0xff]))
-    }
-
-    func testDecoderExposesModelIDAndLabelWithoutChoosing() throws {
-        let blob = antigravityGenerationBlob(
-            model: "gemini-pro-default", input: 10, output: 1, label: " Gemini 3.1 Pro (High) ", timestamp: 1_800_000_000
-        )
-        let event = try XCTUnwrap(AntigravityProtoDecoder.generationEvent(from: blob))
-        XCTAssertEqual(event.modelID, "gemini-pro-default")
-        XCTAssertEqual(event.label, "Gemini 3.1 Pro (High)")
-    }
-
-    func testModelCandidatesPreferLabelsForPlaceholdersAndFoldTieredIDs() {
-        typealias Scanner = AntigravityDbUsageScanner
-        XCTAssertEqual(Scanner.modelCandidates(id: "gemini-default", label: "Gemini 3.5 Flash (Low)"), ["Gemini 3.5 Flash (Low)", "gemini-default"])
-        XCTAssertEqual(Scanner.modelCandidates(id: "gemini-pro-default", label: nil), ["gemini-pro-default"])
-        XCTAssertEqual(Scanner.modelCandidates(id: nil, label: "Claude Opus 4.6 (Thinking)"), ["Claude Opus 4.6 (Thinking)"])
-        XCTAssertEqual(Scanner.modelCandidates(id: "gemini-3.7-flash-high", label: "Gemini 3.7 Flash (High)"), ["gemini-3.7-flash-high", "Gemini 3.7 Flash (High)"])
-        XCTAssertEqual(Scanner.modelCandidates(id: "gemini-3.7-flash-tiered", label: nil), ["gemini-3.7-flash"])
-        XCTAssertEqual(Scanner.modelCandidates(id: nil, label: nil), [])
-    }
-
-    func testSystemPromptOnlyRowsWithoutAnyModelAreBookkeepingNotGenerations() {
-        let bookkeeping = antigravityGenerationBlob(model: nil, input: 0, output: 0, systemPrompt: 1_298, timestamp: 1_800_000_000)
-        let labelled = antigravityGenerationBlob(
-            model: nil, input: 0, output: 0, systemPrompt: 1_298, label: "Gemini 3.1 Pro (Low)", timestamp: 1_800_000_000
-        )
-        let named = antigravityGenerationBlob(model: "gemini-3.6-flash", input: 0, output: 0, systemPrompt: 1_298, timestamp: 1_800_000_000)
-
-        XCTAssertNil(AntigravityProtoDecoder.generationEvent(from: bookkeeping))
-        XCTAssertEqual(AntigravityProtoDecoder.generationEvent(from: labelled)?.inputTokens, 1_298)
-        XCTAssertEqual(AntigravityProtoDecoder.generationEvent(from: named)?.inputTokens, 1_298)
-    }
-
-    func testExtractsStepMetadataTimestamp() {
-        let stepMeta = antigravityStepMetadata(timestamp: 1_800_000_000)
-        XCTAssertEqual(AntigravityProtoDecoder.timestamp(fromStepMetadata: stepMeta), 1_800_000_000)
-        XCTAssertNil(AntigravityProtoDecoder.timestamp(fromStepMetadata: []))
-        XCTAssertNil(AntigravityProtoDecoder.timestamp(fromStepMetadata: [0xff]))
-    }
-
-    func testUsesFallbackTimestampWhenTimingBytesMissing() throws {
-        let blobWithoutTimestamp = antigravityGenerationBlob(
-            model: "gemini-3.8-flash",
-            input: 500,
-            output: 100,
-            timestamp: nil
-        )
-        let event = try XCTUnwrap(AntigravityProtoDecoder.generationEvent(
-            from: blobWithoutTimestamp, stepMetadata: antigravityStepMetadata(timestamp: 1_800_000_123)
-        ))
-        XCTAssertEqual(event.modelID, "gemini-3.8-flash")
-        XCTAssertEqual(event.inputTokens, 500)
-        XCTAssertEqual(event.outputTokens, 100)
-        XCTAssertEqual(event.timestampSeconds, 1_800_000_123)
-    }
-}
-
 final class AntigravityDbUsageScannerTests: XCTestCase {
     private let now = OpenUsageISO8601.date(from: "2026-07-27T12:00:00.000Z")!
     private let pricing = ModelPricing(
@@ -298,25 +71,12 @@ final class AntigravityDbUsageScannerTests: XCTestCase {
 
         let result = await scanner.scan(now: now, pricing: TestPricing.bundled)
         let scan = try XCTUnwrap(result)
+        // Breakdown rows carry no inherent order; the history aggregator sorts them for display.
         let models = try XCTUnwrap(scan.modelUsage?.daily.first?.models)
-        XCTAssertEqual(models.map(\.model), ["gemini-3.1-pro", "gemini-3.7-flash"])
-        XCTAssertEqual(models.map(\.totalTokens), [16, 5])
-    }
-
-    func testDiscoversAntigravityConversationStoresUnderGeminiHome() throws {
-        let home = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-        for path in ["antigravity-cli/conversations", "antigravity-ide", "tmp/conversations", "antigravity/conversations"] {
-            try FileManager.default.createDirectory(
-                at: home.appendingPathComponent(path), withIntermediateDirectories: true
-            )
-        }
-        addTeardownBlock { try? FileManager.default.removeItem(at: home) }
-
         XCTAssertEqual(
-            AntigravityDbUsageScanner.conversationsDirectories(underGeminiHome: home.path),
-            [home.path + "/antigravity/conversations", home.path + "/antigravity-cli/conversations"]
+            Dictionary(uniqueKeysWithValues: models.map { ($0.model, $0.totalTokens) }),
+            ["gemini-3.1-pro": 16, "gemini-3.7-flash": 5]
         )
-        XCTAssertEqual(AntigravityDbUsageScanner.conversationsDirectories(underGeminiHome: "/nonexistent-\(UUID().uuidString)"), [])
     }
 
     func testBookkeepingRowsDoNotRaiseTheUnknownModelWarning() async throws {
@@ -560,48 +320,6 @@ final class AntigravityDbUsageScannerTests: XCTestCase {
         let legacySQL = AntigravityDbUsageScanner.dataSQL(after: 42, includeSteps: false)
         XCTAssertTrue(legacySQL.contains("THEN hex(g.data) ELSE NULL END))"))
         XCTAssertFalse(legacySQL.contains("steps"))
-    }
-
-    /// Reads a real SQLite database through `sqlite3`, so the `steps` correlation and the
-    /// missing-`steps` fallback run against the actual schema.
-    func testEventsWithoutAnyTimestampAreNotAttributedToDatabaseModificationDay() async throws {
-        let fixture = try makeDatabaseDirectory()
-        let path = fixture.paths[0]
-        let sqlite = SQLiteCLIAccessor()
-        let dated = UInt64(now.timeIntervalSince1970) - 3_600
-        let datedHex = antigravityHex(antigravityGenerationBlob(model: "gemini-3.6-flash", input: 10, output: 5, timestamp: dated))
-        let undatedHex = antigravityHex(antigravityGenerationBlob(model: "gemini-3.6-flash", input: 1_000, output: 500, timestamp: nil))
-        try sqlite.execute(path: path, sql: """
-            CREATE TABLE gen_metadata (idx INTEGER PRIMARY KEY, data BLOB);
-            CREATE TABLE steps (idx INTEGER PRIMARY KEY, metadata BLOB);
-            INSERT INTO gen_metadata (idx, data) VALUES (0, X'\(datedHex)'), (1, X'\(undatedHex)');
-            """)
-        let expectedDay = DailyUsageAccumulator.dayKey(from: Date(timeIntervalSince1970: TimeInterval(dated)))
-
-        try FileManager.default.setAttributes([.modificationDate: now], ofItemAtPath: path)
-        let firstScan = await AntigravityDbUsageScanner(
-            sqlite: sqlite, conversationsDirectories: { [fixture.url.path] }
-        ).scan(now: now, pricing: pricing)
-        let first = try XCTUnwrap(firstScan)
-
-        let later = now.addingTimeInterval(86_400)
-        try FileManager.default.setAttributes([.modificationDate: later], ofItemAtPath: path)
-        let secondScan = await AntigravityDbUsageScanner(
-            sqlite: sqlite, conversationsDirectories: { [fixture.url.path] }
-        ).scan(now: later, pricing: pricing)
-        let second = try XCTUnwrap(secondScan)
-
-        XCTAssertEqual(first.series.daily.map(\.date), [expectedDay])
-        XCTAssertEqual(first.series.daily.first?.totalTokens, 15)
-        XCTAssertEqual(second.series.daily, first.series.daily)
-
-        // The same rows without a `steps` table use the legacy query and still skip the undated event.
-        try sqlite.execute(path: path, sql: "DROP TABLE steps")
-        let legacyScan = await AntigravityDbUsageScanner(
-            sqlite: sqlite, conversationsDirectories: { [fixture.url.path] }
-        ).scan(now: later, pricing: pricing)
-        let legacy = try XCTUnwrap(legacyScan)
-        XCTAssertEqual(legacy.series.daily, first.series.daily)
     }
 
     @MainActor
